% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biHazards.R
\name{biHazards}
\alias{biHazards}
\title{Returns non-parametric estimates of the bivariate hazard function}
\usage{
biHazards(formula, data)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms on the right.  The response must be a
survival object as returned by the \code{Surv} function. The RHS must contain a 'cluster' term}

\item{data}{a data.frame in which to interpret the variables named in the
\code{formula} argument.}
}
\value{
List with the estimates of the three bivariate hazard functions. Each hazard function is a matrix.
}
\description{
Returns non-parametric estimates of the bivariate hazard function
}
\seealso{
hazardscpp dabrowska
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
