% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chrCpp}
\alias{chrCpp}
\title{Concordance and discordance matrices for bivariate survival data}
\usage{
chrCpp(x, y, xstatus, ystatus)
}
\arguments{
\item{x, y}{Vectors of failure times}

\item{xstatus, ystatus}{Status indicators for failure times}
}
\value{
Concordance and discordance matrices for bivariate survival data
}
\description{
Concordance and discordance matrices for bivariate survival data
}
\seealso{
CHRtheo
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
