% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrdiff.R
\name{chrdiff}
\alias{chrdiff}
\title{Get integrated squared difference (ISD) between empirical and theoretical CHR}
\usage{
chrdiff(x, y, xstatus, ystatus, par, dist = "gamma", type = "alpha",
  n = 5)
}
\arguments{
\item{x, y}{Vectors of failure times}

\item{xstatus, ystatus}{Status indicators for failure times}

\item{par}{Parameter value for frailty model}

\item{dist}{Frailty distribution. Gamma ("gamma"), positive stable ("posstab") and inverse Gaussian ("invgauss") are available.}

\item{type}{Parameterization of frailty parameter (either "alpha" or "theta")}

\item{n}{How many intervals should the CHR be estimated on.}
}
\value{
Estimated ISD between theoretical and empirical CHR
}
\description{
Get integrated squared difference (ISD) between empirical and theoretical CHR
}
\seealso{
CHRtheo CHR chrCpp
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
