% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{copulaPlot}
\alias{copulaPlot}
\title{Plot of bivariate data, extrapolated and transformed to uniform marginals.}
\usage{
copulaPlot(formula, data = NULL)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms on the right.  The response must be a
survival object as returned by the \code{Surv} function. The RHS must contain a 'cluster' term}

\item{data}{a data.frame containing the variables in the model}
}
\value{
Plot of data, extrapolated, transformed to uniform marginals, and plotted
to compare to copulas.
}
\description{
Plot of bivariate data, extrapolated and transformed to uniform marginals.
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
