% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabrowska.R
\name{dabrowska}
\alias{dabrowska}
\title{Dabrowska estimator of the bivariate survival function}
\usage{
dabrowska(formula, data)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms on the right.  The response must be a
survival object as returned by the \code{Surv} function. The RHS must contain a 'cluster' term}

\item{data}{a data.frame containing the variables in the model}
}
\value{
Matrix with Dabrowska estimate of the bivariate survival function
}
\description{
Dabrowska estimator of the bivariate survival function
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
