% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eyyfunc}
\alias{eyyfunc}
\title{Matrix with expected values for CHR independence test}
\usage{
eyyfunc(x, y, xuni, yuni)
}
\arguments{
\item{x, y}{Vectors of failure times}

\item{xuni, yuni}{Grid points to get expected value for}

\item{xstatus, ystatus}{Status indicators for failure times}
}
\value{
Matrix with expected values for CHR independence test
}
\description{
Matrix with expected values for CHR independence test
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com
}
