% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazards.R
\name{hazards}
\alias{hazards}
\title{Returns non-parametric estimates of the bivariate hazard function}
\usage{
hazards(x, y, xstatus, ystatus)
}
\arguments{
\item{x, y}{Vectors of the failure times}

\item{xstatus, ystatus}{Status indicators for the failure times}
}
\value{
List with the estimates of the three bivariate hazard functions. Each hazard function is a matrix.
}
\description{
Returns non-parametric estimates of the bivariate hazard function
}
\seealso{
hazardscpp dabrowska
}
\author{
Jeppe E. H. Madsen
}
