% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impTailDep.R
\name{impTailDep}
\alias{impTailDep}
\title{Theoretical 'tail dependence' at given quantile for given frailty model with given parameter}
\usage{
impTailDep(q, par, dist = "gamma", type = "alpha", tail = "lwr")
}
\arguments{
\item{q}{Quantile}

\item{par}{Parameter value for frailty model}

\item{dist}{Frailty distribution. Gamma ("gamma"), positive stable ("posstab") and inverse Gaussian ("invgauss") are available.}

\item{type}{Parameterization of frailty parameter (either "alpha" or "theta")}

\item{Tail}{Lower ("lwr") or upper ("upr") tail dependence}
}
\value{
'Tail dependence' at given quantile for given frailty model with given parameter
}
\description{
Theoretical 'tail dependence' at given quantile for given frailty model with given parameter
}
\seealso{
tailDep
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
