% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independenceTest.R
\name{independenceTest}
\alias{independenceTest}
\title{Non-parametric independence test for bivariate survival data}
\usage{
independenceTest(formula, data = NULL, weight = "independence")
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms on the right.  The response must be a
survival object as returned by the \code{Surv} function. The RHS must contain a 'cluster' term}

\item{data}{a data.frame containing the variables in the model}

\item{weight}{weight function for test. Default is 'independence'
which is optimal for the Frank copula. Other options are 'dabrowska', 'atRisk', a weight matrix
or a number (in which case the weight is constant).}
}
\value{
test statistic, SE and p-value for independence test.
}
\description{
Non-parametric independence test for bivariate survival data
}
\examples{
library(survival)
data("diabetic")
independenceTest(Surv(time,status)~cluster(id), data = diabetic)
}
\references{
Shih, Joanna H., and Thomas A. Louis. "Tests of independence for bivariate survival data." Biometrics (1996): 1440-1449.
}
\seealso{
dabrowska
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
