% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHR.R
\name{logLikSort}
\alias{logLikSort}
\title{Sort models according to loglikelihood.}
\usage{
logLikSort(formula, data)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms on the right.  The response must be a
survival object as returned by the \code{Surv} function. The RHS must contain a 'cluster' term.}

\item{data}{a data.frame containing the variables in the model.}
}
\value{
data.frame with loglikelihood for different models, sorted.
}
\description{
Sort models according to loglikelihood
}
\details{
it is mentioned in \code{vignette("frailtyEM_manual")} that the frailty with the
highest loglikelihood should be prefered since all the models are special cases of the PVF frality
family. This is done in a user-friendly way here.
}
\seealso{
CHR autoplot.CHR
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
