% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauCens.R
\name{medianConcordance}
\alias{medianConcordance}
\title{Estimate of median concordance for censored data}
\usage{
medianConcordance(
  formula,
  data = NULL,
  gamma = NULL,
  maxIt = 100,
  method = "dabrowska"
)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms on the right.  The response must be a
survival object as returned by the \code{Surv} function. The RHS must contain a 'cluster' term.}

\item{data}{a data.frame containing the variables in the model.}

\item{gamma}{bandwidth for pruitt estimator. Ignored if estimator is another one.}

\item{maxIt}{maximum number of iterations for NPMLE and pruitt estimators of
bivariate survival function. Doesn't do anything if estimator is 'dabrowska' (default).}

\item{method}{which estimator to use for the bivariate survival function.}
}
\value{
non-parametric estimate of median concordance and parametric estimates from
three different frailty models.
}
\description{
Estimate of median concordance for censored data
}
\details{
median concordance is a rank based measure of dependence which ensures that
it doesn't change if we transform the marginal distributions with strictly increasing functions.
Median concordance is defined as

E(sign((T1 - m1)(T_2 - m2))),

where m1 and m2 are the marginal medians.
Median concordance is more intuitive than for instance Kendall's tau
since you are only comparing whether one pair of observations is concordance or
discordant wrt their respective medians rather than whether one pair is concordant/discordant
compared to another pair. The median concordance is equal to
4*S(m1,m2)-1 where S is the bivariate survival function.
This function estimates S both non-parametrically and parametrically for three frailty models.
The survival function value for the frailty models at the medians are
uniquely determined by the copula they imply since S(m1,m2) = C(S1(m1),S2(m2)) = C(0.5,0.5),
where C is the survival copula implied by the model.
}
\references{
Hougaard, Philip. (2000). Analysis of Multivariate Survival Data.
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
