% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabrowska.R
\name{plot.biSurv}
\alias{plot.biSurv}
\title{Contour plot of the bivariate survival function}
\usage{
\method{plot}{biSurv}(x, ...)
}
\arguments{
\item{x}{an object of class \code{dabrowska}.}

\item{...}{further arguments for \code{ggplot}.}
}
\value{
contour plot of estimate of bivariate survival function
along with contour plot of independence estimate of bivariate survival function.
}
\description{
Contour plot of the Dabrowska estimate of the bivariate survival function
}
\examples{
library(survival)
data("diabetic")
s <- biSurv(Surv(time,status)~cluster(id), data = diabetic)
plot(s)
}
\references{
Hougaard, Philip. (2000). Analysis of Multivariate Survival Data.

Dabrowska, Dorota M. "Kaplan-Meier estimate on the plane." Annals of Statistics 16.4 (1988): 1475-1489.
}
\seealso{
dabrowska print.dabrowska
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
