% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHR.R
\name{plot.CHR}
\alias{plot.CHR}
\title{Plot of CHR as a function of the survival function}
\usage{
\method{plot}{CHR}(x, ...)
}
\arguments{
\item{x}{An object of class \code{CHR}}

\item{...}{Further arguments for \code{ggplot}}
}
\value{
Plot of CHR as a function of the survival function
}
\description{
Plot of CHR as a function of the survival function
}
\references{
Chen, Min-Chi & Bandeen-Roche, Karen. (2005). A Diagnostic for Association in Bivariate Survival Models. Lifetime data analysis. 11. 245-64.
}
\seealso{
chrCpp
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
