% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabrowska.R
\name{plot.dabrowska}
\alias{plot.dabrowska}
\title{Contour plot of the Dabrowska estimate of the bivariate survival function}
\usage{
\method{plot}{dabrowska}(x, ...)
}
\arguments{
\item{x}{An object of class \code{dabrowska}.}

\item{...}{Further arguments for \code{ggplot}.}
}
\value{
Contour plot of bivariate survival function.
}
\description{
Contour plot of the Dabrowska estimate of the bivariate survival function
}
\references{
Chen, Min-Chi & Bandeen-Roche, Karen. (2005). A Diagnostic for Association in Bivariate Survival Models. Lifetime data analysis. 11. 245-64.
}
\seealso{
dabrowska
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
