% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabrowska.R
\name{print.biSurv}
\alias{print.biSurv}
\title{Print a short summary of dabrowska estimate}
\usage{
\method{print}{biSurv}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class "dabrowska"}

\item{digits}{minimal number of _significant_ digits, see \code{print.default}.}

\item{...}{further arguments for \code{ggplot}.}
}
\value{
number of observations, number of events, medians for the marginals,
median concordance, and joint probability of both marginals being greater than respective medians
}
\description{
Print a short summary of dabrowska estimate
}
\examples{
library(survival)
data("diabetic")
s <- biSurv(Surv(time,status)~cluster(id), data = diabetic)
s
plot(s)
}
\references{
Hougaard, Philip. (2000). Analysis of Multivariate Survival Data.

Dabrowska, Dorota M. "Kaplan-Meier estimate on the plane." Annals of Statistics 16.4 (1988): 1475-1489.
}
\seealso{
biSurv plot.biSurv
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
