% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabrowska.R
\name{pruitt}
\alias{pruitt}
\title{Pruitt estimator of bivariate survival function}
\usage{
pruitt(formula, data, gamma, tol = 0.001, maxIt = 25)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
         operator, and the terms on the right.  The response must be a
         survival object as returned by the \code{Surv} function.
The RHS must contain a 'cluster' term}

\item{data}{a data.frame containing the variables in the model}

\item{gamma}{bandwidth of pruitt estimator}

\item{tol}{how large is the difference from iteration to iteration allowed to be}

\item{maxIt}{how many iterations are acceptable}
}
\value{
Matrix with estimate of bivariate survival function
}
\description{
Pruitt estimator of bivariate survival function
}
\seealso{
dabrowska EMsurv
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
