% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFrail.R
\name{simFrail}
\alias{simFrail}
\title{Simulate data from frailty model with Weibull baseline hazard}
\usage{
simFrail(
  n = 100,
  dist = "gamma",
  par = 1,
  shape = 1,
  scale = 1,
  cov1 = 1,
  cov2 = 1,
  beta = 0
)
}
\arguments{
\item{n}{Number of observations in each cluster.}

\item{dist}{Frailty distribution.}

\item{par}{Frailty parameter.}

\item{shape}{shape parameter of Weibull distribution.}

\item{scale}{scale parameter of Weibull distribution.}

\item{cov1}{Covariates for the first vector}

\item{cov2}{Covariates for the second vector}

\item{beta}{Conditional (log-) hazard ratios}
}
\description{
Simulate data from frailty model with Weibull baseline hazard
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
