% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailDep.R
\name{tailDep}
\alias{tailDep}
\title{Estimators of "tail-dependence" for bivariate survival data}
\usage{
tailDep(x, y, xstatus, ystatus, q, tail = "lwr", method = "dabrowska",
  without = NULL)
}
\arguments{
\item{x, y}{Vectors of failure times}

\item{xstatus, ystatus}{Status indicators for failure times}

\item{q}{Quantile to estimate "tail-dependence" for}

\item{tail}{Tail to estimate "tail dependence" for}

\item{method}{What estimator to use}

\item{without}{Distributions to ignore}
}
\value{
Estimate of "tail dependence"
}
\description{
Estimators of "tail dependence" for bivariate survival data
}
\seealso{
tailDepCI
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
