% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailDepCI.R
\name{tailDepCI}
\alias{tailDepCI}
\title{Bootstrapped confidence interval for "tail dependence"}
\usage{
tailDepCI(x, y, xstatus, ystatus, q, method = "dabrowska",
  tail = "lwr", n = 1000)
}
\arguments{
\item{x, y}{Vectors of failure times}

\item{xstatus, ystatus}{Status indicators for failure times}

\item{q}{Quantile to estimate "tail-dependence" for}

\item{method}{How to estimate survival probabilities. Available are 'dabrowska' and 'fast'}

\item{tail}{Tail to estimate "tail dependence" for}

\item{n}{Number of bootstraps}
}
\value{
CI for "tail dependence"
}
\description{
Bootstrapped confidence interval for "tail dependence"
}
\seealso{
tailDep
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
