% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailDependence.R
\name{tailDependence}
\alias{tailDependence}
\title{Estimators of "tail-dependence" for bivariate survival data}
\usage{
tailDependence(formula, data, cluster, q, tail = "lwr",
  method = "dabrowska")
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms on the right.  The response must be a
survival object as returned by the \code{Surv} function. The RHS must contain a 'cluster' term}

\item{data}{a data.frame containing the variables in the model}

\item{cluster}{Cluster variable}

\item{q}{Quantile to estimate "tail-dependence" for}

\item{tail}{Tail to estimate "tail dependence" for}

\item{method}{What estimator to use}
}
\value{
Estimate of "tail dependence"
}
\description{
Estimators of "tail dependence" for bivariate survival data
}
\seealso{
tailDepCI
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
