% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailDependence.R
\name{tailDependencePlot}
\alias{tailDependencePlot}
\title{Function that plots the "tail-dependence" as a function of the chosen quantile}
\usage{
tailDependencePlot(
  formula,
  data,
  q = seq(0.1, 0.2, 0.005),
  tail = "lwr",
  method = "dabrowska"
)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms on the right.  The response must be a
survival object as returned by the \code{Surv} function. The RHS must contain a 'cluster' term}

\item{data}{a data.frame containing the variables in the model}

\item{q}{Quantiles to estimate "tail-dependence" for}

\item{tail}{Tail to estimate "tail dependence" for}

\item{method}{What estimator to use}
}
\value{
Plot of "tail-dependence" as a function of the quantile; estimated and implied by frailty models
}
\description{
Function that plots the "tail-dependence" as a function of the chosen quantile
}
\seealso{
tailDepCI tailDependence
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
