% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{taucpp}
\alias{taucpp}
\title{Returns Kendall's tau and concordance and discordance matrices for bivariate survival data}
\usage{
taucpp(x, y, xstatus, ystatus, KMxsurv, KMysurv, KMxtime, KMytime)
}
\arguments{
\item{x, y}{Vectors of failure times}

\item{xstatus, ystatus}{Status indicators for failure times}

\item{KMxsurv, KMysurv}{Survival functions for the marginal distributions}

\item{KMxtime, KMytime}{Grid points for marginal survival functions}
}
\value{
Kendall's tau and concordance and discordance matrices for bivariate survival data
}
\description{
Returns Kendall's tau and concordance and discordance matrices for bivariate survival data
}
\seealso{
tauCens
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
