% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauCens.R
\name{tauPar}
\alias{tauPar}
\title{Get Kendall's tau from parameter or parameter from Kendall's tau}
\usage{
tauPar(par = 0, dist = "gamma", output = "tau", type = "alpha")
}
\arguments{
\item{par}{parameter value for frailty model.}

\item{dist}{frailty distribution.}

\item{output}{return tau or parameter.}

\item{type}{parameterization of frailty distribution.}
}
\value{
Kendall's tau from parameter or parameter from Kendall's tau.
}
\description{
Get Kendall's tau from parameter or parameter from Kendall's tau
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
