% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniTrans.R
\name{uniTrans}
\alias{uniTrans}
\title{Transform survival data into uniform marginals using a Cox proportional hazards model}
\usage{
uniTrans(formula, data = NULL)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms on the right.  The response must be a
survival object as returned by the \code{Surv} function. The RHS must contain a 'cluster' term}

\item{data}{a data.frame containing the variables in the model}
}
\value{
Data.frame with two vectors of failure times and two vectors of status indicators
}
\description{
Transform survival data into uniform marginals using a Cox proportional hazards model
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
