% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biHazards.R
\name{biHazards}
\alias{biHazards}
\title{Returns non-parametric estimates of the bivariate hazard function}
\usage{
biHazards(x, y, xstatus, ystatus)
}
\arguments{
\item{x, y}{Vectors of failure times}

\item{xstatus, ystatus}{Status indicators for failure times}
}
\value{
List with the estimates of the three bivariate hazard functions. Each hazard function is a matrix.
}
\description{
Returns non-parametric estimates of the bivariate hazard function
}
\seealso{
hazardscpp dabrowska
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
