% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabrowska.R
\name{biSurv}
\alias{biSurv}
\title{Non-parametric estimation of the bivariate survival function}
\usage{
biSurv(
  formula,
  data,
  gamma = NULL,
  tol = 0.001,
  maxIt = 25,
  method = "dabrowska"
)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
         operator, and the terms on the right.  The response must be a
         survival object as returned by the \code{Surv} function.
The RHS must contain a 'cluster' term}

\item{data}{a data.frame containing the variables in the model}

\item{gamma}{bandwidth of pruitt estimator. Can be ignored for other estimators.}

\item{tol}{how large is the difference from iteration to iteration allowed to be}

\item{maxIt}{maximum number of iterations.}

\item{method}{which estimator to use. 'dabrowska' (default), 'pruitt' or NPMLE}
}
\value{
matrix with estimate of bivariate survival function
}
\description{
Non-parametric estimation of the bivariate survival function
}
\details{
TODO
}
\seealso{
print.biSurv plot.biSurv
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
