% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHR.R
\name{CHR}
\alias{CHR}
\title{Integrated squared distance (ISD) between empirical and theoretical CHRs}
\usage{
CHR(x, y, xstatus, ystatus, n = 5, without = NULL)
}
\arguments{
\item{x, y}{Vectors of failure times}

\item{xstatus, ystatus}{Status indicators for failure times}

\item{n}{Number of steps for empirical CHR}

\item{without}{Distributions to ignore}
}
\value{
Data.frame with ISD for different frailty distributions
}
\description{
Integrated squared distance (ISD) between empirical and theoretical CHRs
}
\seealso{
chrCpp
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
