% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHR.R
\name{CHR}
\alias{CHR}
\title{Returns non-parametric estimate of piecewise constant CHR}
\usage{
CHR(x, y, xstatus, ystatus, n = 5)
}
\arguments{
\item{x, y}{Vectors of failure times}

\item{xstatus, ystatus}{Status indicators for failure times}

\item{n}{Number of 'pieces' to estimate CHR on}
}
\value{
Non-parametric estimate of piecewise constant CHR
}
\description{
Returns non-parametric estimate of piecewise constant CHR
}
\seealso{
CHRtheo chrdiff chrCpp
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
