% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHRtheo.R, R/impchr.R
\name{CHRtheo}
\alias{CHRtheo}
\title{Returns theoretical CHR as a function of the survival function for different frailty models}
\usage{
CHRtheo(par, dist = "gamma", type = "alpha")

CHRtheo(par, dist = "gamma", type = "alpha")
}
\arguments{
\item{par}{Parameter value for frailty model}

\item{dist}{Frailty distribution. Gamma ("gamma"), positive stable ("posstab") and inverse Gaussian ("invgauss") are available.}

\item{type}{Parameterization of frailty parameter (either "alpha" or "theta")}

\item{par}{Parameter value for frailty model}

\item{dist}{Frailty distribution. Gamma ("gamma"), positive stable ("posstab") and inverse Gaussian ("invgauss") are available.}
}
\value{
A vector of length 97 with the theoretical CHR when the survival function is equal to 0.01,...,0.97

A vector of length 97 with the theoretical CHR when the survival function is equal to 0.01,...,0.97
}
\description{
Returns theoretical CHR as a function of the survival function for different frailty models

Get theoretical CHR as a function of the survival function for different frailty models
}
\seealso{
chrdiff

chrdiff
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>

Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
