% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabrowska.R
\name{dabrowska}
\alias{dabrowska}
\title{Dabrowska estimator of the bivariate survival function}
\usage{
dabrowska(formula, data)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
         operator, and the terms on the right.  The response must be a
         survival object as returned by the \code{Surv} function.
The RHS must contain a 'cluster' term}

\item{data}{a data.frame containing the variables in the model}
}
\value{
an object of class 'dabrowska' containing the dabrowska estimate of the
bivariate survival function
}
\description{
Dabrowska estimator of the bivariate survival function
}
\details{
the Dabrowska estimator of the bivariate survival function is estimated
and returned along with the independence estimator. This function may be interesting
in and of itself or because of its use in determining goodness of fit internally in the \code{CHR}
function or elsewhere.
}
\examples{
data(kidney)
dabrowska(Surv(time,status) ~ age + sex + disease + cluster(id), data = kidney)
library(survival)
data("diabetic")
s <- dabrowska(Surv(time,status)~cluster(id), data = diabetic)
s
plot(s)
}
\seealso{
print.dabrowska plot.dabrowska
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
