% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabrowska.R
\name{dabrowska}
\alias{dabrowska}
\title{Dabrowska estimator of the bivariate survival function}
\usage{
dabrowska(x, y, xstatus, ystatus)
}
\arguments{
\item{x, y}{Vectors of failure times}

\item{xstatus, ystatus}{Status indicators for failure times}
}
\value{
Matrix with Dabrowska estimate of the bivariate survival function
}
\description{
Dabrowska estimator of the bivariate survival function
}
\seealso{
hazards hazardscpp
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
