% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independenceTest.R
\name{independenceTest}
\alias{independenceTest}
\title{Non-parametric independence test for bivariate survival data}
\usage{
independenceTest(formula, data = NULL, cluster,
  weight = "independence")
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms on the right.  The response must be a
survival object as returned by the \code{Surv} function. The RHS must contain a 'cluster' term}

\item{data}{a data.frame containing the variables in the model}

\item{cluster}{Cluster variable}

\item{weight}{Weight function for test. Default is 'independence' which is optimal for the Frank copula}
}
\value{
Test statistic, SE and p-value for independence test.
}
\description{
Non-parametric independence test for bivariate survival data
}
\seealso{
biHazards
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
