% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailDepCI.R
\name{tailDepCI}
\alias{tailDepCI}
\title{Bootstrapped confidence interval for "tail dependence"}
\usage{
tailDepCI(
  formula,
  data,
  q,
  method = "dabrowska",
  tail = "lwr",
  n = 1000,
  level = 0.95
)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms on the right.  The response must be a
survival object as returned by the \code{Surv} function. The RHS must contain a 'cluster' term}

\item{data}{a data.frame containing the variables in the model}

\item{q}{Quantile to estimate "tail-dependence" for}

\item{method}{How to estimate survival probabilities. Available are 'dabrowska' and 'fast'}

\item{tail}{Tail to estimate "tail dependence" for}

\item{n}{Number of bootstraps}

\item{level}{The confidence level required.}
}
\value{
CI for "tail dependence"
}
\description{
Bootstrapped confidence interval for "tail dependence"
}
\seealso{
tailDep
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
