% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauCens.R
\name{tauCens}
\alias{tauCens}
\title{Estimator of Kendall's tau for censored data}
\usage{
tauCens(formula, data = NULL, cluster, alpha = 0.05,
  method = "adjusted")
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms on the right.  The response must be a
survival object as returned by the \code{Surv} function. The RHS must contain a 'cluster' term}

\item{data}{a data.frame containing the variables in the model}

\item{cluster}{Cluster variable}

\item{alpha}{Significance level}

\item{method}{Which estimator to use}
}
\description{
Estimator of Kendall's tau for censored data
}
\seealso{
tauPar taucpp
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
