% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauCens.R
\name{tauCens}
\alias{tauCens}
\title{Estimator of Kendall's tau for censored data}
\usage{
tauCens(x, y, xstatus, ystatus, alpha = 0.05, method = "adjusted")
}
\arguments{
\item{x, y}{Vectors of failure times}

\item{xstatus, ystatus}{Status indicators for failure times}

\item{alpha}{Significance level}

\item{method}{Which estimator to use}
}
\description{
Estimator of Kendall's tau for censored data
}
\seealso{
tauPar
}
\author{
Jeppe E. H. Madsen <jeppe.ekstrand.halkjaer@gmail.com>
}
